Refactoring package names
Moving utils packages to packages where functions were used or to
package names which better describe functions contained.
