+++ repo/dp/dp_test.go
@@ -1,0 +1,40 @@
+package dp
+
+import (
+	"testing"
+)
+
+func TestSplitOnFirst(t *testing.T) {
+	testDataMatrix := [][]string{
+		{"/usr/bin/share", "/", "", "usr/bin/share"},
+		{"usr/bin/share", "/", "usr", "bin/share"},
+		{"/share", "/", "", "share"},
+		{"share", "/", "share", ""},
+		{"my big testing task", " ", "my", "big testing task"},
+	}
+	for _, testCase := range testDataMatrix {
+		gotPreffix, gotSuffix := splitOnFirst(testCase[0], testCase[1])
+		if gotPreffix != testCase[2] || gotSuffix != testCase[3] {
+			t.Errorf("for SplitOnFirst('%s', '%s'): got ('%s', '%s'), want ('%s', '%s')", testCase[0], testCase[1], gotPreffix, gotSuffix, testCase[2], testCase[3])
+		}
+	}
+}
+
+func TestSplitOnLast(t *testing.T) {
+	testDataMatrix := [][]string{
+		{"/usr/bin/share", "/", "/usr/bin", "share"},
+		{"usr/bin/share", "/", "usr/bin", "share"},
+		{"/share", "/", "", "share"},
+		{"share", "/", "share", ""},
+		{"my big testing task", " ", "my big testing", "task"},
+		{"local:/test1/test2", "/", "local:/test1", "test2"},
+		{"local:/test1", "/", "local:", "test1"},
+		{"local:", "/", "local:", ""},
+	}
+	for _, testCase := range testDataMatrix {
+		gotPreffix, gotSuffix := splitOnLast(testCase[0], testCase[1])
+		if gotPreffix != testCase[2] || gotSuffix != testCase[3] {
+			t.Errorf("for SplitOnLast('%s', '%s'): got ('%s', '%s'), want ('%s', '%s')", testCase[0], testCase[1], gotPreffix, gotSuffix, testCase[2], testCase[3])
+		}
+	}
+}
