+++ utils/errs/errs.go
@@ -1,0 +1,6 @@
+package errs
+
+// Error type is used to create constant errors.
+type Error string
+
+func (e Error) Error() string { return string(e) }
